refactor(#35): move nav into component
